fixes to cylinder and sphere
