+++ inc/parsing.h
@@ -3,6 +3,7 @@
 
 # include "vector.h"
 # include "color.h"
+# include "scene.h"
 
 typedef enum e_id
 {
@@ -31,12 +32,19 @@ typedef struct s_parse
 int	check_valid(t_parse *info);
 int	set_info(t_parse *info);
 
-int create_vector(char *str, t_vec3 *out, double multi);
-int	parse_ambient(char **arr, t_parse *info);
-int	parse_camera(char **arr, t_parse *info);
-int	parse_light(char **arr, t_parse *info);
-int	parse_sphere(char **arr, t_parse *info);
-int	parse_plane(char **arr, t_parse *info);
-int	parse_cylinder(char **arr, t_parse *info);
+int		check_range(t_vec3 *vector, double min, double max);
+int		create_vector(char *str, t_vec3 *out, double multi);
+int		parse_ambient(char **arr, t_parse *info);
+int		parse_camera(char **arr, t_parse *info);
+int		parse_light(char **arr, t_parse *info);
+int		parse_sphere(char **arr, t_parse *info);
+int		parse_plane(char **arr, t_parse *info);
+int		parse_cylinder(char **arr, t_parse *info);
+
+void	init_ambient(t_scene *scene, t_parse info);
+void	init_light(t_scene *scene, t_parse info);
+void	init_plane(t_list **list, t_parse info, int *id);
+void	init_sphere(t_list **list, t_parse info, int *id);
+void	init_cylinder(t_list **list, t_parse info, int *id);
 
 #endif
