+++ scenes/mickey.rt
@@ -1,4 +1,4 @@
-C  0,0,10  0,0,-1  75
+C  0,0,30  0,0,-1  75
 
 L  8.0,2.0,3.0  0.5  240,240,255
 
@@ -11,7 +11,7 @@ cy  5.0,-2.0,4  0.0,1.0,0.0  2.0  3  100,0,255
 
 
 sp  0,0,2  4  0,0,0
-sp  -2,2.5,2  2.5  0,0,0
+sp  -2,2.5,2  2.5  50,50,50
 sp  2,2.5,2  2.5  0,0,0
 
 sp  -4,0,2  1  255,0,0
