+++ testing.rt
@@ -4,7 +4,7 @@ C 0,2,-14         0,-1,-1       60.0
 L 0,5,0 0.5 255,255,255
    
 sp -4.5,-1,0 1 255,0,0
-sp 0,17,-3 20 0,255,0
+sp 0,17,-3 385 0,255,0
 pl 0,-2,0 0,1,0 255,255,255
 cy 0,-1,0 0,1,1 0.5 1.0 255,0,0
 cy 2,-1,0 0,1,0 0.5 2.0 255,0,0
